Restructured folders and files, added ASMDEF files and added namespace to all .cs files
