+++ Editor/com.unity.guid.Editor.asmdef
@@ -1,0 +1,18 @@
+{
+    "name": "com.unity.guid.editor",
+    "rootNamespace": "",
+    "references": [
+        "GUID"
+    ],
+    "includePlatforms": [
+        "Editor"
+    ],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
+}
