+++ Runtime/com.unity.guid.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: bed044abed94ba24e99e357597663950
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
