+++ Samples~/LoadFirst.unity.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: dc5486f439f4c20489bba7901c90ceee
+timeCreated: 1524091778
+licenseType: Pro
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
