+++ Samples~/LoadScenes.cs.meta
@@ -1,5 +1,7 @@
 fileFormatVersion: 2
-guid: dac07e8eb76e66a4fa8c4fe45f186fc4
+guid: f34a8c30583d15e458a12e4388ff1552
+timeCreated: 1524092328
+licenseType: Pro
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
