+++ Samples~/LoadSecond.unity.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 8013f34f0b656684fbc03322c7682ed3
-folderAsset: yes
-timeCreated: 1523920636
+guid: 44b3b16f4b544704d926fd5d5f9b9223
+timeCreated: 1524091841
 licenseType: Pro
 DefaultImporter:
   externalObjects: {}
