+++ Samples~/TestCrossScene.cs.meta
@@ -1,7 +1,5 @@
 fileFormatVersion: 2
-guid: f1c3a067d9be6204c9e067023c1c768b
-timeCreated: 1524092328
-licenseType: Pro
+guid: 9d1d72110525b5441bd3f57fda5df9b4
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
