+++ Samples~/com.unity.guid.samples.asmdef
@@ -1,0 +1,16 @@
+{
+    "name": "com.unity.guid.samples",
+    "rootNamespace": "",
+    "references": [
+        "GUID"
+    ],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
+}
