+++ Samples~/com.unity.guid.samples.asmdef.meta
@@ -1,8 +1,6 @@
 fileFormatVersion: 2
-guid: 6f8ba3c4c3bf19d4999033b50832272b
-timeCreated: 1524091778
-licenseType: Pro
-DefaultImporter:
+guid: 58b9bfc1be0305948a56609d27425d0c
+AssemblyDefinitionImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
