+++ Tests/Runtime/com.unity.guid.Tests.asmdef
@@ -1,0 +1,22 @@
+{
+    "name": "com.unity.guid.Tests",
+    "rootNamespace": "",
+    "references": [
+        "GUID",
+        "UnityEngine.TestRunner",
+        "UnityEditor.TestRunner"
+    ],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": true,
+    "precompiledReferences": [
+        "nunit.framework.dll"
+    ],
+    "autoReferenced": false,
+    "defineConstraints": [
+        "UNITY_INCLUDE_TESTS"
+    ],
+    "versionDefines": [],
+    "noEngineReferences": false
+}
