+++ Tests/Runtime/com.unity.guid.Tests.asmdef.meta
@@ -1,8 +1,6 @@
 fileFormatVersion: 2
-guid: 8acfd1b1a375f2049a2dde40d955ce3f
-timeCreated: 1524091841
-licenseType: Pro
-DefaultImporter:
+guid: af6f01a44c1c3e14ebf0beaf779d5452
+AssemblyDefinitionImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
